# src/main/resources/to-do-openapi.yaml
openapi: 3.0.3
info:
  title: ToDo
  description: ToDo api
  version: 1.0.0
paths:
  /todos:
    get:
      operationId: getToDos
      tags:
        - todos
      description: return all of the todos
      responses:
        '200':
          description: all of the todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToDo'
    post:
      operationId: createToDo
      tags:
        - todos
      description: create a new todo in the database and return it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToDo'
      responses:
        '201':
          description: new todo successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToDo'
  /todos/{id}:
    delete:
      operationId: deleteToDo
      tags:
        - todos
      description: delete todo with provided id from database
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: todo deleted
        '404':
          description: todo not found
    patch:
      operationId: editToDo
      tags:
        - todos
      description: edit todo with provided id to be the object passed in body of request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToDo'
      responses:
        200:
          description: todo edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToDo'
        404:
          description: todo not found


components:
  schemas:
    ToDo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string